+++ firmware/src/controller/robot_commands.h
@@ -10,7 +10,10 @@ void sendArmRobot(uint8_t robot_id);
 void sendEstopRobot(uint8_t robot_id);
 void sendConfirmation(uint8_t robot_id, uint8_t step_id, bool approved);
 void sendStartSequence(uint8_t robot_id, uint8_t sequence_id);
-void sendControlCommand(uint8_t robot_id, int8_t vx, int8_t vy, int8_t omega);
+
+// Updated for RC signals (1000-2000)
+void sendControlCommand(uint8_t robot_id, uint16_t vx, uint16_t vy, uint16_t omega);
+
 void sendDiscover();
 
 #endif // ROBOT_COMMANDS_H
